+++ AmazonChimeSDK/AmazonChimeSDKTests/utils/MediaErrorTests.swift
@@ -1,0 +1,18 @@
+//
+//  MediaErrorTests.swift
+//  AmazonChimeSDK
+//
+//  Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+//  SPDX-License-Identifier: Apache-2.0
+//
+
+@testable import AmazonChimeSDK
+import XCTest
+
+class MediaErrorTests: XCTestCase {
+    func testDescriptionShouldMatch() {
+        XCTAssertEqual(MediaError.audioUninitializedState.description, "audioUninitializedState")
+        XCTAssertEqual(MediaError.audioStartedState.description, "audioStartedState")
+        XCTAssertEqual(MediaError.audioStoppingState.description, "audioStoppingState")
+    }
+}