+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Added `Versioning` class and `sdkVersion` API for retrieving current version of SDK
 - Added `init(frame: CGRect)` initializer in `DefaultVideoRenderView` to properly initialize `DefaultVideoRenderView` in ObjC code.
 - Added basic video features and `MetricsObserver` callback handler in the ObjC demo app.
+- * Added `audioSessionDidDrop` API in `AudioVideoObserver` for temporary disconnects
 
 ### Changed
 - Updated demo app to work with updated [amazon-chime-sdk-js serverless demo](https://github.com/aws/amazon-chime-sdk-js/tree/master/demos/serverless). Note that you need to redeploy the serverless demo to work with the updated demo app